- 클래스 파일 구조화 및 클래스 이름 수정
