+++ app/src/main/java/com/example/mgckiosk/abs/Item.kt
@@ -1,0 +1,4 @@
+package com.example.mgckiosk.abs
+abstract class Item() {
+    abstract fun item()
+}
