+++ app/src/main/java/com/example/mgckiosk/abs/SelectCategory.kt
@@ -1,0 +1,6 @@
+package com.example.mgckiosk.abs
+
+abstract class SelectCategory {
+    abstract var category: Int
+    abstract fun category()
+}
