+++ app/src/main/java/com/example/mgckiosk/abs/SelectNumber.kt
@@ -1,0 +1,6 @@
+package com.example.mgckiosk.abs
+
+abstract class SelectNumber(optionNumber: Int) {
+    abstract fun selectNumber(): Int
+    abstract var optionNum: Int
+}
