+++ app/src/main/java/com/example/mgckiosk/interface/GoBackOrEnd.kt
@@ -1,0 +1,5 @@
+package com.example.mgckiosk.`interface`
+
+interface GoBackOrEnd {
+    fun goBackOrEnd()
+}
