+++ app/src/main/java/com/example/mgckiosk/item/AdeJuice.kt
@@ -1,0 +1,8 @@
+package com.example.mgckiosk.item
+import com.example.mgckiosk.abs.Item
+
+class AdeJuice : Item() {
+    override fun item() {
+        println("에이드 & 커피 리스트")
+    }
+}
