+++ app/src/main/java/com/example/mgckiosk/item/Bread.kt
@@ -1,0 +1,8 @@
+package com.example.mgckiosk.item
+import com.example.mgckiosk.abs.Item
+
+class Bread : Item() {
+    override fun item() {
+        println("빵 리스트")
+    }
+}
