+++ app/src/main/java/com/example/mgckiosk/item/CookieMacaron.kt
@@ -1,0 +1,9 @@
+package com.example.mgckiosk.item
+import com.example.mgckiosk.abs.Item
+
+class CookieMacaron : Item() {
+    override fun item() {
+        println("쿠키 & 마카롱 리스트")
+    }
+
+}
