+++ app/src/main/java/com/example/mgckiosk/item/HotCoffee.kt
@@ -1,0 +1,9 @@
+package com.example.mgckiosk.item
+import com.example.mgckiosk.abs.Item
+
+class HotCoffee : Item() {
+    override fun item() {
+        println("핫 커피 리스트")
+    }
+
+}
