+++ app/src/main/java/com/example/mgckiosk/item/HotTea.kt
@@ -1,0 +1,8 @@
+package com.example.mgckiosk.item
+import com.example.mgckiosk.abs.Item
+
+class HotTea : Item() {
+    override fun item() {
+        println("핫 티 리스트")
+    }
+}
