+++ app/src/main/java/com/example/mgckiosk/item/IcedCoffee.kt
@@ -1,0 +1,9 @@
+package com.example.mgckiosk.item
+
+import com.example.mgckiosk.abs.Item
+
+class IcedCoffee : Item() {
+    override fun item() {
+        println("아이스 커피 리스트")
+    }
+}
