+++ app/src/main/java/com/example/mgckiosk/item/IcedTea.kt
@@ -1,0 +1,8 @@
+package com.example.mgckiosk.item
+import com.example.mgckiosk.abs.Item
+
+class IcedTea : Item() {
+    override fun item() {
+        println("아이스 티 리스트")
+    }
+}
