+++ app/src/main/java/com/example/mgckiosk/item/MugCup.kt
@@ -1,0 +1,10 @@
+package com.example.mgckiosk.item
+
+import com.example.mgckiosk.abs.Item
+
+
+class MugCup : Item() {
+    override fun item() {
+        println("머그컵 리스트")
+    }
+}
