+++ app/src/main/java/com/example/mgckiosk/item/SmoothieFrappe.kt
@@ -1,0 +1,8 @@
+package com.example.mgckiosk.item
+import com.example.mgckiosk.abs.Item
+
+class SmoothieFrappe : Item() {
+    override fun item() {
+        println("스무디 & 프라페 리스트")
+    }
+}
