+++ app/src/main/java/com/example/mgckiosk/item/StickCoffee.kt
@@ -1,0 +1,9 @@
+package com.example.mgckiosk.item
+
+import com.example.mgckiosk.abs.Item
+
+class StickCoffee : Item() {
+    override fun item() {
+        println("스틱 커피 리스트")
+    }
+}
