+++ app/src/main/java/com/example/mgckiosk/item/TeaProduct.kt
@@ -1,0 +1,9 @@
+package com.example.mgckiosk.item
+
+import com.example.mgckiosk.abs.Item
+
+class TeaProduct : Item() {
+    override fun item() {
+        println("티 상품 리스트")
+    }
+}
