+++ app/src/main/java/com/example/mgckiosk/item/Tumbler.kt
@@ -1,0 +1,9 @@
+package com.example.mgckiosk.item
+
+import com.example.mgckiosk.abs.Item
+
+class Tumbler : Item() {
+    override fun item() {
+        println(" 텀블러 상품 리스트")
+    }
+}
